% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue1.R
\name{queue1}
\alias{queue1}
\title{Calculating the Main Quantitative Indicators of Queuing Service System.}
\usage{
queue1(lambda, mu, T)
}
\arguments{
\item{lambda}{The successive arrival intervals of customers to the system are independent and subject to an exponential distribution with a parameter of $\lambda$. (numeric)}

\item{mu}{The service time of the service desk is also iid., and obeys the exponential distribution of parameter $\mu$. (numeric)}

\item{T}{System service time (numeric)}
}
\value{
the length Ls, the average stay time Ws and the probability of customer waiting Pwait.
}
\description{
Queuing problem.
}
\examples{
\dontrun{
lambda <- 4
mu <- 6
T <- 1000
res <- queue1(4,6,1000)
}
}
