% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue2.R
\name{queue2}
\alias{queue2}
\title{Calculating the Main Quantitative Indicators of Queuing Service System.}
\usage{
queue2(lambda, mu, T, S)
}
\arguments{
\item{lambda}{The successive arrival intervals of customers to the system are independent and subject to an exponential distribution with a parameter of $\lambda$. (numeric)}

\item{mu}{The service time of the service desk is also iid., and obeys the exponential distribution of parameter $\mu$. (numeric)}

\item{T}{System service time (numeric)}

\item{S}{The number of the service desk (numeric)}
}
\value{
the length Ls, the average stay time Ws and the probability of customer waiting Pwait.
}
\description{
Queuing problem.
}
\examples{
\dontrun{
lambda <- 4
mu <- 6
T <- 10000
S <- 4
res <- queue2(4,6,10000,4)
}
}
